# model/model_contract.go
/*
 * ODS
 *
 * This is a simple REST API to access Block Chain on Ethereum and handling Smart Contracts and Payment Channel as well.
 *
 * API version: 1.0.0
 * Contact: u.kuehn@tu-berlin.de
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package model

//Contract contains smart contract data to import
type Contract struct {
	Name string `json:"name,omitempty"`

	Abi string `json:"abi,omitempty"`

	Bin string `json:"bin,omitempty"`
}

//Contract contains smart contract data
type ContractsPairPaymentchannel struct {
	Name string `json:"name,omitempty"`

	AssetHolder string `json:"asset_holder,omitempty"`

	Adjudicator string `json:"aajudicator,omitempty"`
}