# model/model_result.go
/*
 * ODS
 *
 * This is a simple REST API to access Block Chain on Ethereum and handling Smart Contracts and Payment Channel as well.
 *
 * API version: 1.0.0
 * Contact: u.kuehn@tu-berlin.de
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package model

//Result does
type Result struct {
	Success bool `json:"success,omitempty"`

	Indata interface{} `json:"indata,omitempty"`

	Outdata interface{} `json:"outdata,omitempty"`

	Exception ModelError `json:"error,omitempty"`

	GasUsage uint64 `json:"gasUsage,omitempty"`
}