# model/model_use_payment_channel.go
/*
 * ODS
 *
 * This is a simple REST API to access Block Chain on Ethereum and handling Smart Contracts and Payment Channel as well.
 *
 * API version: 1.0.0
 * Contact: u.kuehn@tu-berlin.de
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package model

var MessageParts = newMessageTypes()

//UsePaymentChannel does
type UsePaymentChannel struct {
	ReferenceToVersion uint64 `json:"ref-to-version,omitempty"`

	Target string `json:"target,omitempty"`

	Contract string `json:"contract,omitempty"`

	Balance uint64 `json:"balance,omitempty"`

	Part string `json:"part,omitempty"`

	Message string `json:"message,omitempty"`
}

func newMessageTypes() *messageRegistry {
	return &messageRegistry{
		MoneyResponse: "money-response",
		MoneyRequest:  "money-request",
		DataRequest:   "data-request",
		DataResponse:  "data-response",
	}
}

type messageRegistry struct {
	MoneyResponse string
	MoneyRequest  string
	DataRequest   string
	DataResponse  string
}